Force disable unflattened APEX
Android 12 - CAPEX relies on untested features of FS, and well, as one
could expect, it's broken.
Notably, it relies on O_DIRECT over FBE, which is an Android-specific
feature, and which is broken on a number of devices.
Notably, RedMagic 5G won't boot with CAPEX, blaming dm-verity issue on
apex

Thanks @c03d1 for the debug