Fake MTK's ccci_rpcd props
On Mediatek devices, modem can request ro.product.device directly to
switch hardware configuration depending on the HW it is running on.

On AOSP GSI, if the ro.product.device is "generic", it'll switch to
ro.product.vendor.device, but not when it's "phhgsi.*".
Which leads to the device not loading the proper HW config, and failing
to modem altogether (notably on Redmi 9A)

To fix this, have ccci_rpcd report vendor props instead of system props